Module abstutil::abst_paths

Generate paths for different A/B Street files

Structs

MapName
A single map is identified using this. Using a struct makes refactoring later easier, to organize cities hierarchially.

ROOT_DIR

ROOT_PLAYER_DIR

Functions

parse_scenario_path
Extract the map and scenario name from a path. Crashes if the input is strange.

path

path_all_edits

path_all_saves

path_all_scenarios

path_camera_state

path_edits

path_player

path_popdat

path_prebaked_results

path_raw_map

path_save

path_scenario