Module abstutil::io

Functions

find_next_file

find_prev_file
Keeps file extensions

list_all_objects
Just list all things from a directory, return sorted by name, with file extension removed.

load_all_objects
Load all serialized things from a directory, return sorted by name, with file extension removed. Detects JSON or binary. Filters out broken files.

maybe_read_json

read_binary

read_json

read_object