Module abstutil::collections

Structs

Counter

FixedMap
A drop-in replacement for BTreeMap, where the keys have the property of being array indices. Some values may be missing. Much more efficient at operations on individual objects, because it just becomes a simple array lookup.

MultiMap

Tags
Convenience functions around a string->string map

VecMap
Use when your key is just PartialEq, not Ord or Hash.

Traits

IndexableKey
Use with FixedMap. From a particular key, extract a usize. These values should be roughly contiguous; the space used by the FixedMap will be O(n) with respect to the largest value returned here.

Functions

contains_duplicates

retain_btreemap

retain_btreeset

wraparound_get