Struct abstutil::time::Timer

pub struct Timer<'a> {
    results: Vec<String>,
    stack: Vec<StackEntry>,
    outermost_name: String,
    sink: Option<Box<dyn TimerSink + 'a>>,
}

Hierarchial magic

Fields
results: Vec<String>stack: Vec<StackEntry>outermost_name: Stringsink: Option<Box<dyn TimerSink + 'a>>
Implementations

impl<'a> Timer<'a>

pub fn new<S: Into<String>>(raw_name: S) -> Timer<'a>

pub fn new_with_sink(name: &str, sink: Box<dyn TimerSink + 'a>) -> Timer<'a>

pub fn throwaway() -> Timer<'a>

fn temporary_println(&mut self, line: String)

pub fn done(self)

Used to end the scope of a timer early.

pub fn start<S: Into<String>>(&mut self, raw_name: S)

pub fn stop<S: Into<String>>(&mut self, raw_name: S)

pub fn start_iter<S: Into<String>>(&mut self, raw_name: S, total_items: usize)

pub fn next(&mut self)

pub fn cancel_iter_early(&mut self)

pub fn add_result(&mut self, elapsed: f64, line: String)

pub fn parallelize<I, O, F>( &mut self, timer_name: &str, requests: Vec<I>, cb: F ) -> Vec<O>where I: Send, O: Send, F: Send + Clone + Copy + Fn(I) -> O,

Execute the callback over all requests, using all CPUs available. The order of the result is deterministic and matches the input.

pub fn parallelize_polite<I, O, F>( &mut self, timer_name: &str, requests: Vec<I>, cb: F ) -> Vec<O>where I: Send, O: Send, F: Send + Clone + Copy + Fn(I) -> O,

Like parallelize, but leave one CPU free, to avoid thrashing the userโs system.

fn inner_parallelize<I, O, F>( &mut self, timer_name: &str, requests: Vec<I>, cb: F, num_cpus: u32 ) -> Vec<O>where I: Send, O: Send, F: Send + Clone + Copy + Fn(I) -> O,

pub fn retain_parallelized<K, V, F>( &mut self, timer_name: &str, input: BTreeMap<K, V>, keep: F ) -> BTreeMap<K, V>where K: Send + Ord, V: Send, F: Send + Sync + Clone + Copy + Fn(&V) -> bool,

Like BTreeMap::retain, but parallelized

pub fn read_file(&mut self, path: &str) -> Result<()>

Then the caller passes this in as a reader