Struct abstutil::time::Timer

pub struct Timer<'a> {
    results: Vec<String>,
    stack: Vec<StackEntry>,
    outermost_name: String,
    sink: Option<Box<dyn TimerSink + 'a>>,
}

Hierarchial magic

Fields
results: Vec<String>stack: Vec<StackEntry>outermost_name: Stringsink: Option<Box<dyn TimerSink + 'a>>
Implementations

impl<'a> Timer<'a>

pub fn new<S: Into<String>>(raw_name: S) -> Timer<'a>

pub fn new_with_sink(name: &str, sink: Box<dyn TimerSink + 'a>) -> Timer<'a>

pub fn throwaway() -> Timer<'a>

fn println(&mut self, line: String)

fn selfless_println(
maybe_sink: &mut Option<Box<dyn TimerSink + 'a>>,
line: String
)

pub fn done(self)

Used to end the scope of a timer early.

pub fn start<S: Into<String>>(&mut self, raw_name: S)

pub fn stop<S: Into<String>>(&mut self, raw_name: S)

pub fn start_iter<S: Into<String>>(&mut self, raw_name: S, total_items: usize)

pub fn next(&mut self)

pub fn cancel_iter_early(&mut self)

pub fn add_result(&mut self, elapsed: f64, line: String)

pub fn parallelize<I, O, F: Fn(I) -> O>(
&mut self,
timer_name: &str,
requests: Vec<I>,
cb: F
) -> Vec<O> where
I: Send,
O: Send,
F: Send + Clone + Copy,

Execute the callback over all requests, using all CPUs available. The order of the result is deterministic and matches the input.

pub fn parallelize_polite<I, O, F: Fn(I) -> O>(
&mut self,
timer_name: &str,
requests: Vec<I>,
cb: F
) -> Vec<O> where
I: Send,
O: Send,
F: Send + Clone + Copy,

Like parallelize, but leave one CPU free, to avoid thrashing the user’s system.

fn inner_parallelize<I, O, F: Fn(I) -> O>(
&mut self,
timer_name: &str,
requests: Vec<I>,
cb: F,
num_cpus: u32
) -> Vec<O> where
I: Send,
O: Send,
F: Send + Clone + Copy,

pub fn read_file(&mut self, path: &str) -> Result<()>

Then the caller passes this in as a reader