Struct abstutil::time::Timer

pub struct Timer<'a> {
    results: Vec<String>,
    stack: Vec<StackEntry>,
    outermost_name: String,
    notes: Vec<String>,
    pub(crate) warnings: Vec<String>,
    pub(crate) errors: Vec<String>,
    sink: Option<Box<dyn TimerSink + 'a>>,
}

Hierarchial magic

Fields
results: Vec<String>stack: Vec<StackEntry>outermost_name: Stringnotes: Vec<String>warnings: Vec<String>errors: Vec<String>sink: Option<Box<dyn TimerSink + 'a>>
Implementations

impl<'a> Timer<'a>

pub fn new<S: Into<String>>(raw_name: S) -> Timer<'a>

pub fn new_with_sink(name: &str, sink: Box<dyn TimerSink + 'a>) -> Timer<'a>

pub fn throwaway() -> Timer<'a>

fn println(&mut self, line: String)

fn selfless_println(
maybe_sink: &mut Option<Box<dyn TimerSink + 'a>>,
line: String
)

pub fn note(&mut self, line: String)

Log immediately, but also repeat at the end, to avoid having to scroll up and find interesting debug stuff.

pub fn warn(&mut self, line: String)

pub fn error(&mut self, line: String)

pub fn done(self)

Used to end the scope of a timer early.

pub fn start<S: Into<String>>(&mut self, raw_name: S)

pub fn stop<S: Into<String>>(&mut self, raw_name: S)

pub fn start_iter<S: Into<String>>(&mut self, raw_name: S, total_items: usize)

pub fn next(&mut self)

pub fn cancel_iter_early(&mut self)

pub(crate) fn add_result(&mut self, elapsed: f64, line: String)

pub fn parallelize<I, O, F: Fn(I) -> O>(
&mut self,
timer_name: &str,
parallelism: Parallelism,
requests: Vec<I>,
cb: F
) -> Vec<O> where
I: Send,
O: Send,
F: Send + Clone + Copy,

The order of the result is deterministic / matches the input.

pub fn read_file(&mut self, path: &str) -> Result<(), Error>

Then the caller passes this in as a reader